refactor(gui): split modern_gui.py into 6 modules (Phase 4 — move-only)
`gui/modern_gui.py`를 단일 2,405줄 God Object에서 메인 오케스트레이터(~250줄)
+ 4개 탭(`gui/tabs/{recorder,impulcifer,settings,info}_tab.py`) + 다이얼로그
(`gui/dialogs.py`) + 헬퍼(`gui/utils.py`) + 상수(`gui/constants.py`)로 분리.

동작·UI 레이아웃·`generate_brir()` 인자 조립 로직은 한 줄도 변경하지 않은
순수 이동(move-only refactor). 외부 호출처(`gui_main.py`, pyproject 콘솔
스크립트, 테스트, Nuitka 빌드)가 의존하는 `from gui.modern_gui import main_gui`
경로는 그대로 유지된다.

`show_language_selection_dialog`은 60줄짜리 인라인 메서드라 이동 시
인터페이스 설계가 필요하므로 본 PR에서는 `ModernImpulciferGUI`에 잔류시키고
후속 백로그(GitHub 이슈)로 분리한다.

Nuitka는 `--include-package=gui`로 하위 패키지를 재귀 포함하므로 CI 워크플로우
4곳은 변경하지 않았다. `build_scripts/build_nuitka.py`는
`--include-package=gui` 없이 개별 모듈을 나열하는 방식이라 신규 8개 모듈을
명시적으로 추가했다.

검증: py_compile (전체 모듈), ruff (모두 통과), pytest (57 passed, 3 skipped
— GUI 임포트는 headless 환경에서 skip), magnitude_response/virtual_bass/
parallel_processing/통합 테스트 모두 통과, en/ko 271개 i18n 키 동기 유지,
의존성 동기화 확인.
