+++ build_scripts/build_nuitka.py
@@ -166,6 +166,9 @@ def build_impulcifer(project_version="0.0.0", output_base_dir="dist", target_pla
         "core.channel_generation", "core.recorder",
         "core.parallel_processing", "core.parallel_utils", "core.parallel_workers",
         "gui", "gui.modern_gui", "gui.legacy_gui",
+        "gui.constants", "gui.utils", "gui.dialogs",
+        "gui.tabs", "gui.tabs.recorder_tab", "gui.tabs.impulcifer_tab",
+        "gui.tabs.settings_tab", "gui.tabs.info_tab",
         "i18n", "i18n.localization",
         "infra", "infra.logger", "infra.resource_helper", "infra.get_version",
         "infra._build_info",
