+++ gui/tabs/__init__.py
@@ -1,0 +1,1 @@
+# Package marker for gui.tabs
